CarbonFootprintApp/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
openai==0.27.0
psycopg2-binary==2.9.6
